chore(atomix): use separate buffers for readers
The JournalSegmentReaders re-used the MappedByteBuffer from the
JournalSegmentWriter. It may have occured that the Writer closes and
frees up the buffer, but that is was still in use by a reader. By
splitting these up we avoid that.

This is only possible if we guarantee that all events are flushed
onAppend on the follower. To be precise, exactly just before the
follower acknowledges the append as successful.