fix(atomix): map after reading the channel
the brokers didn't start anymore in memory mapped mode. so it might be
necessary to first read the channel, and later map it to a buffer. so
this commit reverts that part

another reason might be that it worked before, becuase the writer wasn't
mapped at all anymore, but now was. So that's a potential later change